Fix missing lzip module warning
This lzip module originates from the original Openarl's PoB code.
I failed to find its sources; anyway, it should be a Windows library.

Making a thin wrapper, for example, for LuaZip, will not be difficult,
but, luckily, lzip is used only for the updating process, which should
be disabled for Nix-controlled installation.